feat(country-tracker): ✨ feature driven
